old/requirements.txt
livekit>=1.0
livekit-agents[silero,turn-detector]~=1.0
livekit-plugins-azure
azure-cognitiveservices-speech
openai>=1.0.0
livekit-agents[openai]~=1.0
websockets
livekit-plugins-noise-cancellation~=0.2
python-dotenv~=1.0
fastapi
uvicorn[standard]
pydantic
httpx
itsdangerous